+++ Friend/firmware/firmware_v1.0/CMakeLists.txt
@@ -13,6 +13,7 @@ target_sources(app PRIVATE
     src/audio.c
     src/codec.c
     src/lib/battery/battery.c
+    src/button.c
 )
 
 target_sources_ifdef(CONFIG_CODEC_OPUS app PRIVATE
