+++ Friend/firmware/firmware_v1.0/src/button.h
@@ -1,0 +1,5 @@
+#pragma once
+
+int button_init();
+void activate_button_work();
+void register_button_service();
