+++ Friend/firmware/firmware_v1.0/src/transport.h
@@ -2,3 +2,4 @@
 #include <zephyr/kernel.h>
 int transport_start();
 int broadcast_audio_packets(uint8_t *buffer, size_t size);
+struct bt_conn *get_current_connection();
